Restructured nodeConfigs & added ThreeJs
