+++ doc/graph.exe/package.json
@@ -22,7 +22,8 @@
     "graph.exe-react": "file:../../../graph.exe-react-0.0.1.tgz",
     "prism-react-renderer": "^1.3.5",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "three": "^0.145.0"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "2.1.0"
