+++ doc/graph.exe/src/NodeConfigs/CalculatorGraph/CustomComp.tsx
