+++ doc/graph.exe/src/NodeConfigs/CalculatorGraph/IO.ts
